# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CareNexus API — Connected Care. Better Outcomes.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clinics:
    get:
      operationId: listClinics
      tags: [clinics]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: areaId
          in: query
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of clinics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClinicList"
    post:
      operationId: createClinic
      tags: [clinics]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClinicInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clinic"
  /clinics/{id}:
    get:
      operationId: getClinic
      tags: [clinics]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Clinic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clinic"
    patch:
      operationId: updateClinic
      tags: [clinics]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClinicUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clinic"
    delete:
      operationId: deleteClinic
      tags: [clinics]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  # ── Roles ─────────────────────────────────────────────────────────────────
...
  /communications:
    get:
      operationId: listCommunications
      tags: [communications]
      security:
        - bearerAuth: []
      parameters:
        - name: patientId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of communications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunicationList"
    post:
      operationId: createCommunication
      tags: [communications]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommunicationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Communication"
  /communications/{id}:
    get:
      operationId: getCommunication
      tags: [communications]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Communication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Communication"
    delete:
      operationId: deleteCommunication
      tags: [communications]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  # ── Notifications ─────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
    Meta:
      type: object
      required: [total, page, limit, pages]
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        pages:
          type: integer

    # ── Auth ────────────────────────────────────────────────────────────────
...
    Area:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Clinic:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        createdAt:
          type: string
        area:
          $ref: "#/components/schemas/Area"
    ClinicInput:
      type: object
      required: [name, areaId]
      properties:
        name:
          type: string
        areaId:
          type: string
        address:
          type: string
        city:
          type: string
        phone:
          type: string
        email:
          type: string
    ClinicUpdate:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        city:
          type: string
        phone:
          type: string
        email:
          type: string
    ClinicList:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Clinic"
        meta:
          $ref: "#/components/schemas/Meta"

    # ── Role ────────────────────────────────────────────────────────────────
...
    Communication:
      type: object
      required: [id, patientId, type, subject, status, createdAt]
      properties:
        id:
          type: string
        patientId:
          type: string
        type:
          type: string
        subject:
          type: string
        body:
          type: ["string", "null"]
        status:
          type: string
        sentAt:
          type: ["string", "null"]
        createdAt:
          type: string
    CommunicationInput:
      type: object
      required: [patientId, type, subject]
      properties:
        patientId:
          type: string
        type:
          type: string
        subject:
          type: string
        body:
          type: string
    CommunicationList:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Communication"
        meta:
          $ref: "#/components/schemas/Meta"

    # ── Notification ────────────────────────────────────────────────────────
...